fix: fixtures and updatePatientPictograms were fixed

+++ src/services/patientPictograms.service.js
@@ -746,7 +746,7 @@ module.exports = {
 
         // on the other hand, if the imageUrl from the patientPictograms is not the default or is different from the pictogram imageUrl we wiil update the image
         // and delete the old one.
-        if(patientPictogramExist.imageUrl !== defaultPictogramImage || pictogramExist.imageUrl !== patientPictogramExist.imageUrl) {
+        if(patientPictogramExist.imageUrl !== defaultPictogramImage && pictogramExist.imageUrl !== patientPictogramExist.imageUrl) {
           const { url, ...restResponse}  = await azureImages.updateAndUploadImage(file, imageName, pictogramContainer);
           handleError = restResponse;
 
@@ -945,8 +945,7 @@ module.exports = {
 
       // Delete patientPictogram image in the azure container only if the image is not the default one or if the image
       // belongs to General Pictograms
-      if(patientPictogramExist.imageUrl !== pictogramExist.imageUrl) {
-        if(patientPictogramExist.imageUrl !== defaultPictogramImage) {
+      if(patientPictogramExist.imageUrl !== pictogramExist.imageUrl && patientPictogramExist.imageUrl !== defaultPictogramImage) {
         const imageName = patientPictogramExist.imageUrl.split('/').pop();
         const { error, statusCode, message } = await azureImages.deleteAzureImage(imageName, pictogramContainer);
 
@@ -958,7 +957,6 @@ module.exports = {
             message
           }
         }
-      }
       }
 
       // Commit transaction
